Se arreglo el bug que permitia borrar temas sin sesion abierta en dashboard.js, tambien se agrego filtro a metodo borrar de temasController.php y se modificoconsulta SQL en Tema.php metodo delete para mejorar la seguridad al borrar tema

+++ Controllers/temasController.php
@@ -50,11 +50,17 @@
 		}
 
 		public function borrar(){
-			$this->tema->set('id', trim(strip_tags($_POST['idte'])));
-			$this->tema->set('titulo', trim(strip_tags($_POST['tite'])));
+			if(isset($_SESSION['userSesion']) && $_SESSION['userSesion']['rol'] == 'Tutor'){
+				$this->tema->set('id', trim(strip_tags($_POST['idte'])));
+				$this->tema->set('titulo', trim(strip_tags($_POST['tite'])));
+				$this->tema->set('id_usuario', trim(strip_tags($_POST['idus'])));
 
-			$return = $this->tema->delete();
-			echo $return;
+				$return = $this->tema->delete();
+				echo $return;
+			}
+			else{
+				header("Location:".URL);
+			}
 		}
 
 		public function paginar_temas($arg){
